pyproject.toml
[tool.poetry]
name = "isoai"
version = "0.1.0"
description = "AI PROJECT FOR ISTANBUL CHAMBER OF INDUSTRY"
authors = ["gorkemkaramolla <gorkem_karamolla@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
flask = "^3.0.3"
opencv-python = "^4.9.0.80"
onnxruntime = "^1.17.3"
onnx = "^1.16.0"
scikit-image = "^0.23.2"
transformers = "^4.40.2"
torch = "^2.3.0"
flask-cors = "^4.0.1"
pyannote-audio = "^3.2.0"
pydub = "^0.25.1"
torchvision = "^0.18.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}


oracledb = "^2.2.0"
python-dotenv = "^1.0.1"
logging = "^0.4.9.6"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
isoai = "isoai:main"